Created reusable custom alert component
